requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.4.2
openai==1.3.0
prometheus-client==0.17.1
cachetools==5.3.2
transformers==4.36.2
torch>=2.2.0
accelerate==0.26.1
sentencepiece==0.1.99
pytest==8.0.0
pytest-asyncio==0.23.5 